Beta 6.2

+++ client/src/pages/login/Login.page.jsx
@@ -32,7 +32,11 @@ const LoginPage = () => {
         let databases = await window.indexedDB.databases();
         if(databases.length > 0) {
             databases.forEach((database, index) => {
-                window.indexedDB.deleteDatabase(database.name)
+                if(database.name === '3Ds' || database.name === 'MachinesParts') {
+
+                }else{
+                    window.indexedDB.deleteDatabase(database.name)
+                }
             })
         }
     }
